making all pieces more accessible and adding trader class
