+++ typescript/phoenix-sdk/src/events.ts
@@ -5,6 +5,7 @@ import {
 } from "@solana/web3.js";
 import { BinaryReader } from "borsh";
 import base58 from "bs58";
+import BN from "bn.js";
 
 import { PROGRAM_ID } from "./index";
 import {
@@ -102,7 +103,7 @@ export async function getEventsFromTransaction(
           tradeEvents.push({
             index: reader.readU16(),
             makerId: readPublicKey(reader),
-            orderSequenceNumber: reader.readU64(),
+            orderSequenceNumber: new BN(reader.readFixedArray(8)),
             priceInTicks: reader.readU64(),
             baseLotsFilled: reader.readU64(),
             baseLotsRemaining: reader.readU64(),
@@ -111,7 +112,7 @@ export async function getEventsFromTransaction(
         case PhoenixMarketEvent.Place:
           tradeEvents.push({
             index: reader.readU16(),
-            orderSequenceNumber: reader.readU64(),
+            orderSequenceNumber: new BN(reader.readFixedArray(8)),
             clientOrderId: reader.readU64(),
             priceInTicks: reader.readU64(),
             baseLotsPlaced: reader.readU64(),
