+++ typescript/phoenix-sdk/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from "./numbers";
+export * from "./market";
